Fixes #1612
git-svn-id: https://buddypress.svn.wordpress.org/trunk@2386 cdf35c40-ae34-48e0-9cc9-0c9da1808c22

+++ bp-themes/bp-sn-parent/profile/change-avatar.php
@@ -12,62 +12,70 @@
 
 		<?php do_action( 'bp_before_profile_avatar_upload_content' ) ?>
 
-		<p><?php _e( 'Your avatar will be used on your profile and throughout the site. If there is a <a href="http://gravatar.com">Gravatar</a> associated with your account email we will use that, or you can upload an image from your computer.', 'buddypress') ?></p>
+		<?php if ( !(int)get_site_option( 'bp-disable-avatar-uploads' ) ) : ?>
 
-		<form action="" method="post" id="avatar-upload-form" enctype="multipart/form-data">
+			<p><?php _e( 'Your avatar will be used on your profile and throughout the site. If there is a <a href="http://gravatar.com">Gravatar</a> associated with your account email we will use that, or you can upload an image from your computer.', 'buddypress') ?></p>
 
-			<?php if ( 'upload-image' == bp_get_avatar_admin_step() ) : ?>
+			<form action="" method="post" id="avatar-upload-form" enctype="multipart/form-data">
 
-				<h3><?php _e( 'Your Current Avatar', 'buddypress' ) ?></h3>
+				<?php if ( 'upload-image' == bp_get_avatar_admin_step() ) : ?>
 
-				<p id="current-avatar">
-					<?php bp_displayed_user_avatar( 'type=full') ?>
-					<?php bp_displayed_user_avatar( 'type=thumb' ) ?>
-				</p>
+					<h3><?php _e( 'Your Current Avatar', 'buddypress' ) ?></h3>
 
-				<?php if ( bp_get_user_has_avatar() ) : ?>
-					<div class="generic-button" id="delete-avatar-button">
-						<a class="edit" href="<?php bp_avatar_delete_link() ?>" title="<?php _e( 'Delete Avatar', 'buddypress' ) ?>"><?php _e( 'Delete Avatar', 'buddypress' ) ?></a>
-					</div>
+					<p id="current-avatar">
+						<?php bp_displayed_user_avatar( 'type=full') ?>
+						<?php bp_displayed_user_avatar( 'type=thumb' ) ?>
+					</p>
+
+					<?php if ( bp_get_user_has_avatar() ) : ?>
+						<div class="generic-button" id="delete-avatar-button">
+							<a class="edit" href="<?php bp_avatar_delete_link() ?>" title="<?php _e( 'Delete Avatar', 'buddypress' ) ?>"><?php _e( 'Delete Avatar', 'buddypress' ) ?></a>
+						</div>
+					<?php endif; ?>
+
+					<h3><?php _e( 'Upload a New Avatar', 'buddypress' ) ?></h3>
+
+					<p><?php _e( 'Click below to select a JPG, GIF or PNG format photo from your computer and then click \'Upload Image\' to proceed.', 'buddypress' ) ?></p>
+
+					<p id="avatar-upload">
+						<input type="file" name="file" id="file" />
+						<input type="submit" name="upload" id="upload" value="<?php _e( 'Upload Image', 'buddypress' ) ?>" />
+						<input type="hidden" name="action" id="action" value="bp_avatar_upload" />
+					</p>
+
+					<?php wp_nonce_field( 'bp_avatar_upload' ) ?>
+
 				<?php endif; ?>
 
-				<h3><?php _e( 'Upload a New Avatar', 'buddypress' ) ?></h3>
+				<?php if ( 'crop-image' == bp_get_avatar_admin_step() ) : ?>
 
-				<p><?php _e( 'Click below to select a JPG, GIF or PNG format photo from your computer and then click \'Upload Image\' to proceed.', 'buddypress' ) ?></p>
+					<h3><?php _e( 'Crop Your New Avatar', 'buddypress' ) ?></h3>
 
-				<p id="avatar-upload">
-					<input type="file" name="file" id="file" />
-					<input type="submit" name="upload" id="upload" value="<?php _e( 'Upload Image', 'buddypress' ) ?>" />
-					<input type="hidden" name="action" id="action" value="bp_avatar_upload" />
-				</p>
+					<img src="<?php bp_avatar_to_crop() ?>" id="avatar-to-crop" class="avatar" alt="<?php _e( 'Avatar to crop', 'buddypress' ) ?>" />
 
-				<?php wp_nonce_field( 'bp_avatar_upload' ) ?>
+					<div id="avatar-crop-pane">
+						<img src="<?php bp_avatar_to_crop() ?>" id="avatar-crop-preview" class="avatar" alt="<?php _e( 'Avatar preview', 'buddypress' ) ?>" />
+					</div>
 
-			<?php endif; ?>
+					<input type="submit" name="avatar-crop-submit" id="avatar-crop-submit" value="<?php _e( 'Crop Image', 'buddypress' ) ?>" />
 
-			<?php if ( 'crop-image' == bp_get_avatar_admin_step() ) : ?>
+					<input type="hidden" name="image_src" id="image_src" value="<?php bp_avatar_to_crop_src() ?>" />
+					<input type="hidden" id="x" name="x" />
+					<input type="hidden" id="y" name="y" />
+					<input type="hidden" id="w" name="w" />
+					<input type="hidden" id="h" name="h" />
 
-				<h3><?php _e( 'Crop Your New Avatar', 'buddypress' ) ?></h3>
+					<?php wp_nonce_field( 'bp_avatar_cropstore' ) ?>
 
-				<img src="<?php bp_avatar_to_crop() ?>" id="avatar-to-crop" class="avatar" alt="<?php _e( 'Avatar to crop', 'buddypress' ) ?>" />
+				<?php endif; ?>
 
-				<div id="avatar-crop-pane">
-					<img src="<?php bp_avatar_to_crop() ?>" id="avatar-crop-preview" class="avatar" alt="<?php _e( 'Avatar preview', 'buddypress' ) ?>" />
-				</div>
+			</form>
 
-				<input type="submit" name="avatar-crop-submit" id="avatar-crop-submit" value="<?php _e( 'Crop Image', 'buddypress' ) ?>" />
+		<?php else : ?>
 
-				<input type="hidden" name="image_src" id="image_src" value="<?php bp_avatar_to_crop_src() ?>" />
-				<input type="hidden" id="x" name="x" />
-				<input type="hidden" id="y" name="y" />
-				<input type="hidden" id="w" name="w" />
-				<input type="hidden" id="h" name="h" />
+			<p><?php _e( 'Your avatar will be used on your profile and throughout the site. To change your avatar, please create an account with <a href="http://gravatar.com">Gravatar</a> using the same email address as you used to register with this site.', 'buddypress' ) ?></p>
 
-				<?php wp_nonce_field( 'bp_avatar_cropstore' ) ?>
-
-			<?php endif; ?>
-
-		</form>
+		<?php endif; ?>
 
 		<?php do_action( 'bp_after_profile_avatar_upload_content' ) ?>
